% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnalyzeEntitiesResponse}
\alias{AnalyzeEntitiesResponse}
\title{AnalyzeEntitiesResponse Object}
\usage{
AnalyzeEntitiesResponse(entities = NULL, language = NULL)
}
\arguments{
\item{entities}{The recognized entities in the input document}

\item{language}{The language of the text, which will be the same as the language specified}
}
\value{
AnalyzeEntitiesResponse object
}
\description{
AnalyzeEntitiesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The entity analysis response message.
}
