% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnalyzeSyntaxResponse}
\alias{AnalyzeSyntaxResponse}
\title{Google Cloud Natural Language API Objects 
Google Cloud Natural Language API provides natural language understanding technologies to developers. Examples include sentiment analysis, entity recognition, and text annotations.}
\usage{
AnalyzeSyntaxResponse(language = NULL, sentences = NULL, tokens = NULL)
}
\arguments{
\item{language}{The language of the text, which will be the same as the language specified}

\item{sentences}{Sentences in the input document}

\item{tokens}{Tokens, along with their syntactic information, in the input document}
}
\value{
AnalyzeSyntaxResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:00:11
filename: /Users/mark/dev/R/autoGoogleAPI/googlelanguagev1.auto/R/language_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AnalyzeSyntaxResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The syntax analysis response message.
}
