% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnnotateTextRequest}
\alias{AnnotateTextRequest}
\title{AnnotateTextRequest Object}
\usage{
AnnotateTextRequest(encodingType = NULL, document = NULL, features = NULL)
}
\arguments{
\item{encodingType}{The encoding type used by the API to calculate offsets}

\item{document}{Input document}

\item{features}{The enabled features}
}
\value{
AnnotateTextRequest object
}
\description{
AnnotateTextRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for the text annotation API, which can perform multipleanalysis types (sentiment, entities, and syntax) in one call.
}
\seealso{
Other AnnotateTextRequest functions: \code{\link{documents.annotateText}}
}
