% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{AnnotateTextResponse}
\alias{AnnotateTextResponse}
\title{AnnotateTextResponse Object}
\usage{
AnnotateTextResponse(language = NULL, sentences = NULL, tokens = NULL,
  entities = NULL, documentSentiment = NULL)
}
\arguments{
\item{language}{The language of the text, which will be the same as the language specified}

\item{sentences}{Sentences in the input document}

\item{tokens}{Tokens, along with their syntactic information, in the input document}

\item{entities}{Entities, along with their semantic information, in the input document}

\item{documentSentiment}{The overall sentiment for the document}
}
\value{
AnnotateTextResponse object
}
\description{
AnnotateTextResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The text annotations response message.
}
