% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Document}
\alias{Document}
\title{Document Object}
\usage{
Document(gcsContentUri = NULL, language = NULL, type = NULL,
  content = NULL)
}
\arguments{
\item{gcsContentUri}{The Google Cloud Storage URI where the file content is located}

\item{language}{The language of the document (if not specified, the language is}

\item{type}{Required}

\item{content}{The content of the input in string format}
}
\value{
Document object
}
\description{
Document Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
################################################################ #Represents the input to API methods.
}
