% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_functions.R
\name{documents.analyzeEntities}
\alias{documents.analyzeEntities}
\title{Finds named entities (currently finds proper names) in the text,entity types, salience, mentions for each entity, and other properties.}
\usage{
documents.analyzeEntities(AnalyzeEntitiesRequest)
}
\arguments{
\item{AnalyzeEntitiesRequest}{The \link{AnalyzeEntitiesRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/natural-language/}{Google Documentation}

Other AnalyzeEntitiesRequest functions: \code{\link{AnalyzeEntitiesRequest}}
}
