% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_functions.R
\name{documents.analyzeSyntax}
\alias{documents.analyzeSyntax}
\title{Analyzes the syntax of the text and provides sentence boundaries andtokenization along with part of speech tags, dependency trees, and otherproperties.}
\usage{
documents.analyzeSyntax(AnalyzeSyntaxRequest)
}
\arguments{
\item{AnalyzeSyntaxRequest}{The \link{AnalyzeSyntaxRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/natural-language/}{Google Documentation}

Other AnalyzeSyntaxRequest functions: \code{\link{AnalyzeSyntaxRequest}}
}
