% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_functions.R
\name{documents.annotateText}
\alias{documents.annotateText}
\title{A convenience method that provides all the features that analyzeSentiment,analyzeEntities, and analyzeSyntax provide in one call.}
\usage{
documents.annotateText(AnnotateTextRequest)
}
\arguments{
\item{AnnotateTextRequest}{The \link{AnnotateTextRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/natural-language/}{Google Documentation}

Other AnnotateTextRequest functions: \code{\link{AnnotateTextRequest}}
}
