% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Entity.metadata}
\alias{Entity.metadata}
\title{Entity.metadata Object}
\usage{
Entity.metadata()
}
\value{
Entity.metadata object
}
\description{
Entity.metadata Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Metadata associated with the entity.Currently, Wikipedia URLs and Knowledge Graph MIDs are provided, ifavailable. The associated keys are 'wikipedia_url' and 'mid', respectively.
}
\seealso{
Other Entity functions: \code{\link{Entity}}
}
