% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{EntityMention}
\alias{EntityMention}
\title{EntityMention Object}
\usage{
EntityMention(text = NULL, type = NULL)
}
\arguments{
\item{text}{The mention text}

\item{type}{The type of the entity mention}
}
\value{
EntityMention object
}
\description{
EntityMention Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a mention for an entity in the text. Currently, proper nounmentions are supported.
}
