% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_functions.R
\docType{package}
\name{language_googleAuthR}
\alias{language_googleAuthR}
\alias{language_googleAuthR-package}
\title{Google Cloud Natural Language API
Google Cloud Natural Language API provides natural language understanding technologies to developers. Examples include sentiment analysis, entity recognition, and text annotations.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:00:11
filename: /Users/mark/dev/R/autoGoogleAPI/googlelanguagev1.auto/R/language_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
}
}
