% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Sentence}
\alias{Sentence}
\title{Sentence Object}
\usage{
Sentence(text = NULL, sentiment = NULL)
}
\arguments{
\item{text}{The sentence text}

\item{sentiment}{For calls to AnalyzeSentiment or if}
}
\value{
Sentence object
}
\description{
Sentence Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a sentence in the input document.
}
