% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{TextSpan}
\alias{TextSpan}
\title{TextSpan Object}
\usage{
TextSpan(beginOffset = NULL, content = NULL)
}
\arguments{
\item{beginOffset}{The API calculates the beginning offset of the content in the original}

\item{content}{The content of the output text}
}
\value{
TextSpan object
}
\description{
TextSpan Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents an output piece of text.
}
