% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language_objects.R
\name{Token}
\alias{Token}
\title{Token Object}
\usage{
Token(text = NULL, dependencyEdge = NULL, lemma = NULL,
  partOfSpeech = NULL)
}
\arguments{
\item{text}{The token text}

\item{dependencyEdge}{Dependency tree parse for this token}

\item{lemma}{[Lemma](https://en}

\item{partOfSpeech}{Parts of speech tag for this token}
}
\value{
Token object
}
\description{
Token Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the smallest syntactic building block of the text.
}
